+++ ProjectA/src/main/java/com/example/ProjectA/repository/RolePermissionRepository.java
@@ -4,5 +4,5 @@ import com.example.ProjectA.entity.RolePermission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RolePermissionRepository extends JpaRepository<RolePermission, Long> {
-    boolean existsByRoleIdAndPermissionId(int roleId, int permissionId);
+    boolean existsByRoleIdAndPermissionId(Long roleId, Long permissionId);
 }
